Add simple concurrency to Microhttp. Split event loop into (1) single passive-server-socket event loop and (2) many active-socket events loops. Each event loop is single-threaded and independent. Still no locking or synchronization. This design is facilitated by Selector, which permits select on one thread and register on another.
